Implement listening from micher and from microphone

+++ src/main/java/ua/ttsagent/audio/VoiceHandlerImpl.java
@@ -1,6 +1,7 @@
 package ua.ttsagent.audio;
 
 import jakarta.validation.constraints.NotEmpty;
+import javafx.scene.control.TextArea;
 import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
@@ -11,12 +12,16 @@ import ua.ttsagent.config.FileOutputConfig;
 
 import javax.sound.sampled.AudioSystem;
 import javax.sound.sampled.DataLine;
+import javax.sound.sampled.Mixer;
 import javax.sound.sampled.TargetDataLine;
 import java.io.File;
+import java.io.IOException;
 import java.time.ZonedDateTime;
 import java.util.Arrays;
+import java.util.LinkedList;
 import java.util.UUID;
 
+import static ua.ttsagent.audio.HandlerAction.START;
 import static ua.ttsagent.audio.HandlerAction.STOP;
 
 @Component
@@ -27,10 +32,12 @@ class VoiceHandlerImpl implements VoiceHandler {
     private final AudioFormatConfig audioFormatConfig;
     private final FileOutputConfig fileOutputConfig;
     private volatile HandlerAction action = STOP;
+    private LinkedList<Voice> voiceList = new LinkedList<>();
+
 
     @SneakyThrows
     @Override
-    public File startHandleVoice(@NotEmpty String input) {
+    public void startHandleVoice(@NotEmpty String input) {
         var audioFormat = audioFormatConfig.getAudioFormat();
         var inputInfo = Arrays.stream(AudioSystem.getMixerInfo())
                 .filter(info -> info.getName().equalsIgnoreCase(input.toLowerCase()))
@@ -46,8 +53,9 @@ class VoiceHandlerImpl implements VoiceHandler {
         log.info("Started capturing audio from {}", input);
         byte[] buffer = new byte[4096];
 
-        var resultFile = new File(fileOutputConfig.getOutputDir().toString() + "/" + UUID.randomUUID()+"out.wav");
+        File resultFile = new File(fileOutputConfig.getOutputDir().toString() + "/" + UUID.randomUUID() + "out.wav");
         StreamingWavWriter writer = null;
+        action = START;
         try {
             if (resultFile.exists()) {
                 resultFile.delete();
@@ -55,22 +63,49 @@ class VoiceHandlerImpl implements VoiceHandler {
             resultFile.createNewFile();
             writer = new StreamingWavWriter(resultFile, 44100, 16, 1);
             var timeout = ZonedDateTime.now().plus(3, java.time.temporal.ChronoUnit.SECONDS);
-            while (!ZonedDateTime.now().isAfter(timeout)) {
+            voiceList.push(new Voice(resultFile, targetLine, inputMixer, writer));
+            while (action != STOP) {
                 targetLine.read(buffer, 0, buffer.length);
                 writer.writeData(buffer);
             }
+        } catch (IOException e) {
+            log.info("Stream writer is closed, file {} is deleted", resultFile.getName());
         } catch (Exception e) {
             log.error("Error creating wav file", e);
-        } finally {
-            if (writer != null) {
-                log.info("Closing wav writer");
-                writer.close();
-                targetLine.close();
-                inputMixer.close();
-                resultFile.deleteOnExit();
-            }
         }
-        return resultFile;
     }
 
+
+    @Override
+    public File stopsHandleVoice(TextArea outputArea) {
+        action = STOP;
+        Voice voice = voiceList.poll();
+        voiceList.clear();
+        if (voice != null) {
+
+            var resultFile = voice.resultFile();
+            try {
+                var writer = voice.writer();
+                var targetDataLine = voice.targetDataLine();
+                var inputMixer = voice.inputMixer();
+                if (writer != null) {
+                    writer.close();
+                    targetDataLine.close();
+                    inputMixer.close();
+                    resultFile.deleteOnExit();
+                    log.info("Wav writer is closed, file {} is deleted", resultFile.getName());
+                }
+            } catch (Exception e) {
+                outputArea.setText("Error something wrong please try again later.");
+                log.error("Error creating wav file", e);
+            }
+
+            return resultFile;
+        }
+        return null;
+    }
+
+    record Voice(File resultFile, TargetDataLine targetDataLine,
+                 Mixer inputMixer, StreamingWavWriter writer) {
+    }
 }
